+++ src/module-ag-grid-community.d.ts
@@ -1,0 +1,17 @@
+/**
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+import { CrossValidationOptions } from './components/spreadsheet/utils/equipment-table-utils';
+
+declare module 'ag-grid-community' {
+    // used to add properties that are not supported by ColDef such as numeric, fractionDigits...
+    interface ColDef {
+        numeric?: boolean;
+        fractionDigits?: number;
+        crossValidation?: CrossValidationOptions;
+    }
+}
