+++ src/module-commons-ui.d.ts
@@ -53,7 +53,9 @@ declare module '@gridsuite/commons-ui' {
         previousValue?: string;
         allowNewValue?: boolean;
         onChangeCallback?: () => void;
-        getOptionLabel?: (rt: string | { id: string; label: string }) => string | null;
+        getOptionLabel?: (
+            rt: string | { id: string; label: string }
+        ) => string | null;
         formProps?: Omit<
             TextFieldProps,
             'value' | 'onChange' | 'inputRef' | 'inputProps' | 'InputProps'
@@ -185,7 +187,7 @@ declare module '@gridsuite/commons-ui' {
 
     export const DARK_THEME: string;
     export const LIGHT_THEME: string;
-  
+
     interface CheckboxInputProps {
         name: string;
         label?: string;
