+++ tsconfig.json
@@ -17,5 +17,5 @@
         "noEmit": true,
         "jsx": "react-jsx"
     },
-    "include": ["src","decs.d.ts"]
+    "include": ["src"]
 }
